Hey, a sudoku solver
This has had a few local revisions while I got to grips with, er, not
being a boob with C. But I sorted out the last remaining bug (it was not
checking a dependent cell for index 0 ‐ see line 77!) while watching
Fnatic win the grand final for CS:GO at IEM Katowice. So here it is.